% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierInspect.R
\name{outlierInspect}
\alias{outlierInspect}
\title{Outlier inspection produces 4 graphs}
\usage{
outlierInspect(data)
}
\arguments{
\item{data}{The data that is being inspected}
}
\description{
Outlier inspection produces 4 graphs
}
\examples{
outlierInspect(data)



}
